Rename the SchemaLookup component to SchemaPicker
The Vue picker component shared its name with the unrelated SchemaLookup
data-access interface (application/SchemaLookup.ts) — public-api.ts even
re-exported both under the name SchemaLookup (one as a type, one as a
value). Rename the component, its placeholder message, and its CSS class
to SchemaPicker so the UI picker and the data-access lookup are clearly
distinct. The data interface keeps the SchemaLookup name, paired with
SchemaRepository.
